fix(attribute-converter): fix default type for generic attribute converter from Component to any
With the default type of the AttributeConverter being Component, a converter method can't be run
from within an arrow function (in tests, e.g.) as the implied this-type of the arrow function
doesn't match Component. This change only modifies the default typing, not the fact that the
component instance will call converter methods with the component instance as this-context.