refactor(overlay): add overlay-config mixin defining all overlay config properties on a component
The overlay component can be configured using an overlay-config object. For greater flexibility, we
want to expose each of the overlay config properties as properties on the overlay component itself.
In addition, we want to trigger a component update, each time one of the overlay config properties
change. For this we need a host of setters and getters to synchronize each of the properties with
the overall configuration object and to trigger an update request in case a change happens. To keep
the overlay component clean, we can use a class mixin for the overlay config properties which we
mixin to the overlay component.